+++ backend/problems/src/ma1/arithmetics/order_of_operations.rs
@@ -19,7 +19,7 @@ fn addition_multiplication(id: i32, lang: Language) -> Result<Problem> {
     let solution_text = get_solution(id, lang)?;
     let solution = format!(
         "{solution_text} \\
-        ${term_1} + colored({factor_1} dot {factor_2}) = {term_1} + colored({product}) = {answer}$
+        $ {term_1} + colored({factor_1} dot {factor_2}) = {term_1} + colored({product}) = {answer} $
         "
     );
     Ok(Problem {
@@ -48,7 +48,7 @@ fn subtraction_multiplication(id: i32, lang: Language) -> Result<Problem> {
     let solution_text = get_solution(id, lang)?;
     let solution = format!(
         "{solution_text} \\
-        ${term_1} - colored({factor_1} dot {factor_2}) = {term_1} - colored({product}) = {answer}$
+        $ {term_1} - colored({factor_1} dot {factor_2}) = {term_1} - colored({product}) = {answer} $
         "
     );
     Ok(Problem {
@@ -78,8 +78,8 @@ fn mult_add_mult(id: i32, lang: Language) -> Result<Problem> {
     let solution_text = get_solution(id, lang)?;
     let solution = format!(
         "{solution_text} \\
-        $colored({factor_1} dot {factor_2}) + colored({factor_3} dot {factor_4}) =
-        colored({product_1}) + colored({product_2}) = {answer}$"
+        $ colored({factor_1} dot {factor_2}) + colored({factor_3} dot {factor_4}) =
+        colored({product_1}) + colored({product_2}) = {answer} $"
     );
     Ok(Problem {
         id,
@@ -111,8 +111,8 @@ fn mult_sub_mult(id: i32, lang: Language) -> Result<Problem> {
     let solution_text = get_solution(id, lang)?;
     let solution = format!(
         "{solution_text} \\
-        $colored({factor_1} dot {factor_2}) - colored({factor_3} dot {factor_4}) =
-        colored({product_1}) - colored({product_2}) = {answer}$"
+        $ colored({factor_1} dot {factor_2}) - colored({factor_3} dot {factor_4}) =
+        colored({product_1}) - colored({product_2}) = {answer} $"
     );
     Ok(Problem {
         id,
@@ -121,5 +121,37 @@ fn mult_sub_mult(id: i32, lang: Language) -> Result<Problem> {
         solution,
         identifiers: vec![factor_1, factor_2, factor_3, factor_4],
         combinations: f_range.len().pow(2),
+    })
+}
+
+/// 10 + (2 + 1) * 3
+/// Absolute difficulty: 1
+/// Relative difficulty: 3
+#[problem]
+fn add_par_mult(id: i32, lang: Language) -> Result<Problem> {
+    let (term_1, t_range) = num_gen::integer().range(1, 5).and_random();
+    // The numbers inside the parentheses
+    let term_2 = t_range.random();
+    let term_3 = t_range.random();
+    let (factor, f_range) = num_gen::integer().range(2, 10).and_random();
+
+    let par_sum = term_2 + term_3;
+    let product = par_sum * factor;
+    let answer = term_1 + product;
+
+    let question = format!("${term_1} + ({term_2} + {term_3}) dot {factor}$");
+    let solution_text = get_solution(id, lang)?;
+    let solution = format!(
+        "{solution_text} \\
+        $ {term_1} + colored(({term_2} + {term_3})) dot {factor} = {term_1} + colored({par_sum} dot {factor}) = \\
+        = {term_1} + {product} = {answer} $"
+    );
+    Ok(Problem {
+        id,
+        question,
+        answer: answer.as_math(),
+        solution,
+        identifiers: vec![term_1, term_2, term_3, factor],
+        combinations: f_range.len() * t_range.len().pow(3),
     })
 }